Add cloud function entry point
This is adding the main entry point for the Google Cloud Functions
Framework. It's basically a single event handling function which gets
called whenever a file finishes uploading in the specified GCS bucket.

The cloud function reads the config from a YAML config file that is
expected to be present in the repo (follows in a subsequent PR).

The YAML config file stores configs for multiple cloud functions. Which
one to use is communicated via an environment variable `config_name`.

Note that the entry point file has to be called `main.py`.