Add DB import config for comparative benchmarks
This is adding a pipeline config for the comparative benchmarking
results.

Most of the complexity in the commit comes from the SQL epxressions that
are part of the config for unit testing. They are the same which we also
use in the database and they need to support a variety of old data
schemas that we have in the database.