Move InMemoryStorage into separate file
The tests for the `rules` was using an in-memory implementation of the
cloud storage client API. I'm moving this into a separate file to be
able to use that from other tests as well.

The implementation is now also complete in the sense that it is
implementing the previously established storage protocols.

It is also adding the `prefix` parameter to the list_blobs function in
the protocol and in both implementations (`LocalStorage` and
`InMemoryStorage`).