Add batch import tool to DB import
This is adding a new tool to the CLI, the batch import.

The batch import can scan through all the files in an existing
GCS bucket and trigger the import pipeline as if all those
files were just uploaded. It's usually needed when a new
pipeline is set up and when we want to import all the existing data.