Add verify subcommand to CLI tool
The `verify` command allows testing the whole pipeline locally without
the need for a BigQuery database server or temporary table, or without
the need for deploying a cloud function.

It uses the bigquery-emulator to spin up a local in memory database that
behaves like BigQuery. It then trigger the cloud function processing
code for a set of predefined files that already exist in the source
bucket. Afterwards it executes a series of predefined SQL queries to
check whether all the data has been correctly imported into the
database.