Add download helper tool
This is adding a small helper tool to the CLI which allows the user to
download all the needed files from a bucket to their local hard drive.

The tool accepts the name of a pipeline configuration as an argument,
then gets a list of all the files from the bucket associated with the
config. Each file is treated as a trigger to the pipeline, similar to
how the batch import works.

The user will end up with a local copy of all the files that are needed
to trigger the cloud function locally.

In the example of IREE:
- Benchmark traces would NOT be downloaded because the config leaves
  those alone
- benchmark-results-*.json files will be downloaded because these are
  valid triggers of the pipeline
- execution-benchmark-config.json files will ALSO be downloaded because
  the data transformation step will request those as needed even though
  they won't trigger an import on its own.