Add deployment tool to DB Import
This adds a deploy command to the CLI tool.

The deploy command can deploy a cloud function.
It reads the config from config file and takes care of all the steps:

1. Creating the cloud function.
2. Creating the trigger.
3. Authorizing the cloud function's service account to write to the
   destination bigquery table.
4. Authorizing the trigger to call the cloud function.
5. Authorizing the bucket's service account to trigger the trigger.
6. (Optional) Create the table
7. (Optional) Delete old data already in a table.
8. (Optional) Perform a batch import of files already existing in the
   source bucket.

It does all these steps the simple way by calling `glcoud` and `bq`. I
think it would be possible to do all of that using only REST APIs but
that would be a lot harder to implement.

That also means that the deploy command needs `gcloud` to be properly
set up and authenticated. If not the tool will just fail with some
gcloud errors.