Rename cloud_functions config key to pipelines
This renames the main config key in the config file from
`cloud_functions` to `pipelines` because this will make the usage of
terms in the documentation more consistent.

It's because each entry under `pipelines` has the configuration of the
full assembly, not just the cloud function. The cloud function is just
one aspect of it, that why the docs call it a pipeline.