Use anonymous credentials in bigquery emulator
This is a quick fixup for the bigquery emulator. It explicitly makes the
database client use anonymous credentials - which is all we need when
connecting to the emulator.

If this was not the case then the client library would try to infer
credentials from the user's gcloud config and fail if the user didn't
have one.