[DB Import] Replace bq mk by SQL in deploy command (#71)
The deploy command can automatically create the destination table in
bigquery if it didn't exist beforehand. So far it was using the bq
command line tool for that.

This change replaces the bq call by a SQL command because we will need
the SQL command anyway for automatically creating tables with the
bigquery emulator.